refactor(http): share minimal atomic file helpers
